clk: qcom: Support bypass RCG configuration
In the case of HDMI clocks, we want to bypass the RCG's ability
to divide the output clock and pass through the parent HDMI PLL
rate through. Add a simple set of clk_ops to configure the RCG to
do this. This removes the need to keep adding more frequency
entries to the tv_src clock whenever we want to support a new
rate.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>