drm: Lobotomize set_busid nonsense for platform drivers
To be able to nerf the set_busid callback no-op it out like udl
already does. But that means GET_UNIQUE will return an empty string,
and someone (mesa drivers) might actually care about that. Hence use
dev->unique as a fallback in that ioctl if it's there. This is also
the fallback in the set_busid function, but doing it that way will
allow us to not touch all the drivers, at least for now.

v2: Make drm_dev_set_unique static and update kerneldoc.

Cc: Emil Velikov <emil.l.velikov@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>