drm: add exclusive resource handling interfaces.
The idea here is for VR compositors.

There is an opt in interface for modesetting masters (X, compositors)
that they can state they will respect the allowed resources list,
and will update it on hotplug events.

This list contains crtcs, planes and connectors the modesetting
master will interact with. Any modesetting operations on these
objects will be ignored by the kernel. The list will change on
a hotplug event, and needs to be re-read in full.

Then there is another root-only ioctl. The is to be used by
 the VR process compositor (possbly via a suid-root helper or
logind). This ioctl takes a list of objects for exclusive mode
and tags them as exclusive for that fd, it also disconnects
the fd from the current master and makes it into a disconnected
master of it's own.

TODO:
DRM_MASTER still blocks this from working, as we want
a non-current master to use modesetting ioctls.